Repair production builds by replacing categories generator
