+++ Rakefile
@@ -10,7 +10,6 @@ desc 'Builds the site locally'
 task :build do
   puts 'Building site.'
   sh 'PRODUCTION="YES" jekyll build --destination _gh-pages'
-  sh 'PRODUCTION="YES" jekyll build --destination _gh-pages'
 end
 
 namespace :podcast do
