+++ _layouts/category_list_page.html
@@ -10,10 +10,10 @@ footer: false
   <article>
     <p>We covered a lot of topics in the last few years, here's all of the tags we've applied to our posts.</p>
     <ul>
-    {% for post_category in page.post_categories %}
-    <a title="{{ post_category.count }} posts on {{ post_category.name }}" href="/{{ post_category.dir }}">{{ post_category.name }}</a>,
+    {% for category in page.categories %}
+      <li><a title="{{ category.count }} posts on {{ category.name }}" href="/blog/categories/{{ category.dir }}">{{ category.name }}</a></li>
     {% endfor %}
     </ul>
-    </article>
+  </article>
 </div>
