+++ _layouts/epic.html
@@ -60,7 +60,7 @@
 
           <section>
             <h4>Author</h4>
-            <p>{% for key in page.author %}{% assign author = site.authors[key] %}{% if forloop.index > 1 %},{% endif %} <a href="/author/{{key}}"/>{{ author.name }}</a>{% endfor %}</p>
+            <p>{% for key in page.author %}{% assign author = site.authors[key] %}{% if forloop.index > 1 %},{% endif %} <a href="/author/{{key}}">{{ author.name }}</a>{% endfor %}</p>
             <p>{% for key in page.author %}{% assign author = site.authors[key] %}{{ author.byline }}</a>{% endfor %}</p>
           </section>
